folder setup update

+++ 10_folder_setup.sh
@@ -24,8 +24,10 @@ mkdir -p ~/${SUBDIR}/telegraf
 
 cd ~/${SUBDIR}
 ##find . -type d -exec touch {}/.gitignore \;
-rm ~/${SUBDIR}/mosquitto/log/.gitignore
-
+if [ -e ~/${SUBDIR}/mosquitto/log/.gitignore ]
+then
+    rm ~/${SUBDIR}/mosquitto/log/.gitignore
+fi
 
 touch ~/${SUBDIR}/mosquitto/config/mosquitto.conf
 touch ~/${SUBDIR}/mosquitto/config/passwd
@@ -33,8 +35,9 @@ touch ~/${SUBDIR}/mosquitto/log/mosquitto.log
 touch ~/${SUBDIR}/influxdb/influxdb.conf
 touch ~/${SUBDIR}/telegraf/telegraf.conf
 sudo setfacl -Rdm g:docker:rwx ~/${SUBDIR}
-sudo chmod -R 775 ~/${SUBDIR}
 sudo chmod -R ug+rw ~/${SUBDIR}
 sudo chmod -R o+r ~/${SUBDIR}
+sudo chmod -R 664 ~/${SUBDIR}
+sudo chmod -R 775 ~/${SUBDIR}/*.sh
 sudo chown -R ${USER}:docker ~/${SUBDIR}
